+++ tests/autofix_files/async913_trio_anyio.py
@@ -1,0 +1,12 @@
+# ARG --enable=ASYNC913
+# AUTOFIX
+# ASYNCIO_NO_ERROR
+
+import trio
+
+
+async def nursery_no_cancel_point():
+    while True:  # ASYNC913: 4
+        await trio.lowlevel.checkpoint()
+        async with trio.open_nursery():
+            pass